+++ packages/drizzle/schemas/bts.ts
@@ -298,6 +298,8 @@ export const gsmCells = pgTable(
     createdAt: timestamp({ withTimezone: true }).notNull().defaultNow(),
   },
   (t) => [
+    check("gsm_lac_check", sql`${t.lac} BETWEEN 0 AND 65535`),
+    check("gsm_cid_check", sql`${t.cid} BETWEEN 0 AND 65535`),
     unique("gsm_cells_lac_cid_unique").on(t.cell_id, t.lac, t.cid),
     index("gsm_cells_cid_idx").on(t.cid),
     index("gsm_cells_cid_trgm_idx").using("gin", sql`(${t.cid}::text) gin_trgm_ops`),
@@ -322,6 +324,10 @@ export const umtsCells = pgTable(
     createdAt: timestamp({ withTimezone: true }).notNull().defaultNow(),
   },
   (t) => [
+    check("umts_lac_check", sql`${t.lac} BETWEEN 0 AND 65535`),
+    check("umts_rnc_check", sql`${t.rnc} BETWEEN 0 AND 65535`),
+    check("umts_cid_check", sql`${t.cid} BETWEEN 0 AND 65535`),
+    check("umts_arfcn_check", sql`${t.arfcn} BETWEEN 0 AND 16383`),
     unique("umts_cells_rnc_cid_unique").on(t.cell_id, t.rnc, t.cid),
     index("umts_cells_cid_idx").on(t.cid),
     index("umts_cells_cid_trgm_idx").using("gin", sql`(${t.cid}::text) gin_trgm_ops`),
@@ -350,7 +356,9 @@ export const lteCells = pgTable(
   },
   (t) => [
     check("clid_check", sql`${t.clid} BETWEEN 0 AND 255`),
+    check("lte_enbid_check", sql`${t.enbid} BETWEEN 0 AND 1048575`),
     check("pci_check", sql`${t.pci} BETWEEN 0 AND 503`),
+    check("lte_earfcn_check", sql`${t.earfcn} BETWEEN 0 AND 262143`),
     unique("lte_cells_enbid_clid_unique").on(t.cell_id, t.enbid, t.clid),
     index("lte_cells_iot_true_idx")
       .on(t.enbid, t.clid)
@@ -382,6 +390,11 @@ export const nrCells = pgTable(
     createdAt: timestamp({ withTimezone: true }).notNull().defaultNow(),
   },
   (t) => [
+    check("nr_nrtac_check", sql`${t.nrtac} BETWEEN 0 AND 16777215`),
+    check("nr_gnbid_check", sql`${t.gnbid} BETWEEN 0 AND 4294967295`),
+    check("nr_clid_check", sql`${t.clid} BETWEEN 0 AND 16383`),
+    check("nr_pci_check", sql`${t.pci} BETWEEN 0 AND 1007`),
+    check("nr_arfcn_check", sql`${t.arfcn} BETWEEN 0 AND 3279165`),
     // unique("nr_cells_gnbid_clid_unique").on(t.gnbid, t.clid).nullsNotDistinct(),
     index("nr_cells_redcap_true_idx")
       .on(t.gnbid, t.clid)